int-to-double fixes & sophia update

+++ ProgrammingModels/ThetaGPU/CUDA/vecadd_mpi/main.cpp
@@ -15,8 +15,8 @@ int main( int argc, char* argv[] )
 
   double result_ = offload(me,nranks);
 
-  int sum;
-  MPI_Reduce(&result_, &sum, 1, MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+  double sum;
+  MPI_Reduce(&result_, &sum, 1, MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
 
   if(me == 0) {
     if(sum < 1e-6) printf("\nResult is CORRECT!! :)\n");
